requirements.txt
Django>=3.0
psycopg2-binary>=2.8
termplotlib
djangorestframework
markdown
django-filter
django-cors-headers
